pyproject.toml
[project]
name = "stock-research-assistant"
version = "1.0.0"
description = "AI-Powered Stock Research Assistant for Indian Markets using CrewAI and Mistral AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # CrewAI Framework
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    
    # LLM Integration
    "litellm>=1.40.0",
    
    # Telegram Bot
    "python-telegram-bot>=21.0",
    
    # Web Scraping
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # Indian Market Data
    "yfinance>=0.2.40",
    "nsetools>=1.0.11",
    
    # Technical Analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    
    # Data Processing
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    
    # Web UI
    "streamlit>=1.30.0",
    
    # Web UI
    "streamlit>=1.32.0",
    
    # Database (optional)
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "bot", "crews", "tools"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
